Allow `swifttemplate` files to "import" Swift files when building the template.
Summary:
As per discussion in https://github.com/krzysztofzablocki/Sourcery/pull/663,
the initial idea of inlining parts of the `.swifttemplate` files into the intermediate `main.swift` file at the top level
to allow protocol declarations and extensions was rejected because of the number of reasons.

A better approach would be to allow `.swifttemplate` file to reference other Swift files,
so that when the template is being built, the referenced files are getting compiled into a single module.

In this commit, a support for `<%- import("fileName[.swift]") %>` commands in `.swifttemplate` files is added.
File extension could be omitted, in which case ".swift" extension is used.
"Imported" files are just plain normal Swift files that are being compiled by `swiftc`.
The files are copied to the temporary build folder, their sha256 is used to determine if the contents has changed.